refactor(discovery): add dedicated DiscoveryError in error.rs
